Move payload and hearbeat related stuff to an internal pkg
